analytica/requirements.txt
fastapi
uvicorn
python-jose
psycopg2-binary
pandas
openai
plotly
jinja2
